requirements.txt
uvicorn
fastapi
slowapi
pyjwt
cffi
sqlalchemy[asyncio]
aiosqlite
python-dotenv
aiohttp
loguru
cryptography
python-jose[cryptography]
httpx
psutil